updated hooks-output and hooks-output-utils
