--- package/hooks_generator/hooks_src/output/hooks-outputlayout.ts
@@ -1,2 +1,0 @@
-// import tpye { Ledger } from '../hook-types.ts';
-// import ____ './hooks-output.ts'
